[STYLE] : 컴포넌트 분리도 해봤습니다
